# cmd/server/swagger/openapi3.json
{"components":{"requestBodies":{"PostGiteaWebhookRequestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GiteaPayload"}}},"description":"Request body for Gitea webhook"}},"responses":{"ErrorResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Status"}}},"description":"Something went wrong!"},"HealthCheckResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Status"}}},"description":"Server is running!"},"WebhookAcceptedResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Status"}}},"description":"Server is running!"}},"schemas":{"GiteaPayload":{"properties":{"action":{"type":"string"},"issue":{"properties":{"body":{"type":"string"},"title":{"type":"string"},"url":{"type":"string"},"user":{"properties":{"email":{"type":"string"}},"type":"object"}},"type":"object"},"number":{"format":"int64","type":"integer"},"repository":{"properties":{"name":{"type":"string"},"owner":{"properties":{"login":{"type":"string"}},"type":"object"}},"type":"object"}},"type":"object"},"Status":{"properties":{"message":{"default":"Some message from the backend!","type":"string"}},"type":"object"}},"securitySchemes":{"basicAuth":{"description":"HTTP Basic authentication","scheme":"basic","type":"http"}}},"info":{"contact":{"email":"5030347+b1tray3r@users.noreply.github.com","name":"aborgardt","url":"https://github.com/b1tray3r/gitea-issues-to-redmine"},"description":"This API provides endpoints to sync Gitea issues with redmine.","title":"Gitea-Issues-To-Redmine-Sync","version":"1.0.0"},"openapi":"3.0.3","paths":{"/gitea/webhook":{"description":"Takes in the webhook content from Gitea","post":{"requestBody":{"$ref":"#/components/requestBodies/PostGiteaWebhookRequestBody"},"responses":{"200":{"$ref":"#/components/responses/WebhookAcceptedResponse"}},"summary":"Endpoint filled by a Gitea wehook.","tags":["Gitea"]},"summary":"Webhook endpoint for Gitea."},"/health":{"description":"Checks the availablility of the API server.","get":{"responses":{"200":{"$ref":"#/components/responses/HealthCheckResponse"}},"summary":"Get health status","tags":["General"]},"summary":"HealthCheck without authentication"}},"servers":[{"description":"localhost","url":"http://localhost:8085"}]}