# pkg/api/openapi3.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// GiteaPayload defines model for GiteaPayload.
type GiteaPayload struct {
	Action *string `json:"action,omitempty"`
	Issue  *struct {
		Body  *string `json:"body,omitempty"`
		Title *string `json:"title,omitempty"`
		Url   *string `json:"url,omitempty"`
		User  *struct {
			Email *string `json:"email,omitempty"`
		} `json:"user,omitempty"`
	} `json:"issue,omitempty"`
	Number     *int64 `json:"number,omitempty"`
	Repository *struct {
		Name  *string `json:"name,omitempty"`
		Owner *struct {
			Login *string `json:"login,omitempty"`
		} `json:"owner,omitempty"`
	} `json:"repository,omitempty"`
}

// Status defines model for Status.
type Status struct {
	Message *string `json:"message,omitempty"`
}

// HealthCheckResponse defines model for HealthCheckResponse.
type HealthCheckResponse = Status

// WebhookAcceptedResponse defines model for WebhookAcceptedResponse.
type WebhookAcceptedResponse = Status

// PostGiteaWebhookRequestBody defines model for PostGiteaWebhookRequestBody.
type PostGiteaWebhookRequestBody = GiteaPayload

// PostGiteaWebhookJSONRequestBody defines body for PostGiteaWebhook for application/json ContentType.
type PostGiteaWebhookJSONRequestBody = GiteaPayload

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Endpoint filled by a Gitea wehook.
	// (POST /gitea/webhook)
	PostGiteaWebhook(ctx echo.Context) error
	// Get health status
	// (GET /health)
	GetHealth(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostGiteaWebhook converts echo context to params.
func (w *ServerInterfaceWrapper) PostGiteaWebhook(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostGiteaWebhook(ctx)
	return err
}

// GetHealth converts echo context to params.
func (w *ServerInterfaceWrapper) GetHealth(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetHealth(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/gitea/webhook", wrapper.PostGiteaWebhook)
	router.GET(baseURL+"/health", wrapper.GetHealth)

}

type HealthCheckResponseJSONResponse Status

type WebhookAcceptedResponseJSONResponse Status

type PostGiteaWebhookRequestObject struct {
	Body *PostGiteaWebhookJSONRequestBody
}

type PostGiteaWebhookResponseObject interface {
	VisitPostGiteaWebhookResponse(w http.ResponseWriter) error
}

type PostGiteaWebhook200JSONResponse struct {
	WebhookAcceptedResponseJSONResponse
}

func (response PostGiteaWebhook200JSONResponse) VisitPostGiteaWebhookResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetHealthRequestObject struct {
}

type GetHealthResponseObject interface {
	VisitGetHealthResponse(w http.ResponseWriter) error
}

type GetHealth200JSONResponse struct {
	HealthCheckResponseJSONResponse
}

func (response GetHealth200JSONResponse) VisitGetHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Endpoint filled by a Gitea wehook.
	// (POST /gitea/webhook)
	PostGiteaWebhook(ctx context.Context, request PostGiteaWebhookRequestObject) (PostGiteaWebhookResponseObject, error)
	// Get health status
	// (GET /health)
	GetHealth(ctx context.Context, request GetHealthRequestObject) (GetHealthResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PostGiteaWebhook operation middleware
func (sh *strictHandler) PostGiteaWebhook(ctx echo.Context) error {
	var request PostGiteaWebhookRequestObject

	var body PostGiteaWebhookJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostGiteaWebhook(ctx.Request().Context(), request.(PostGiteaWebhookRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostGiteaWebhook")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostGiteaWebhookResponseObject); ok {
		return validResponse.VisitPostGiteaWebhookResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetHealth operation middleware
func (sh *strictHandler) GetHealth(ctx echo.Context) error {
	var request GetHealthRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealth(ctx.Request().Context(), request.(GetHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealth")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetHealthResponseObject); ok {
		return validResponse.VisitGetHealthResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}